archive/blog: add old html blog entries
converted with pandoc then corrected by hand. There are some deadlinks
to https://getkiss.org/ - should they be changed? Some of those pointed
to pages which we might now have in this wiki, some might be on the new
KISS website. Should the ones with new versions be updated or should
they all be left?